migrate and seed tested working

+++ data/migrations/01-make_cars_table.js
@@ -5,7 +5,7 @@ exports.up = function (knex) {
     tbl.string('vin', 17).notNullable().unique()
     tbl.string('make', 130).notNullable()
     tbl.string('model', 130).notNullable()
-    tbl.numeric('mileage').unsigned().notNullable()
+    tbl.integer('mileage').unsigned().notNullable()
     tbl.string('title', 130)
     tbl.string('transmission', 130)
   })
